fix: Fix enum for server connection and GUI updates
The enum for trying to connect to the server has been fixed, resolving any issues related to the connection state in the GUI. Additionally, the enum has been updated to announce the state of the connection in the GUI. This includes renaming the enum to better reflect whether we are connected to the server or not. The merge conflict in the GUI, specifically in the menu, has also been resolved. Other modifications include updating the position of the error message when the client fails to connect to the server and adding a new private member in the game menu to store the state of the game. Finally, the core event handling has been fixed.